address feedback provided by @jamesbirdsall
